# openapi.yaml
openapi: 3.1.0
info:
  title: SpecR API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >
    Headless REST API for CSI® MasterFormat®-compatible specification document
    automation with round-trip DOCX support. Independent project; not affiliated
    with CSI.

# No authentication yet — auth lands with Phase 5f (#43). The explicit empty
# security array documents that every endpoint is currently unauthenticated.
security: []
servers:
  - url: http://localhost:3000
    description: Local development server

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /specs/{id}:
    get:
      operationId: getSpec
      summary: Retrieve a spec with its full paragraph tree
      tags: [specs]
      parameters:
        - $ref: '#/components/parameters/SpecId'
      responses:
        '200':
          description: Spec with paragraph tree
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/SpecTree'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      operationId: updateSpec
      summary: Update spec metadata
      tags: [specs]
      parameters:
        - $ref: '#/components/parameters/SpecId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                section:
                  type: string
                  description: CSI section number (e.g. "09 91 26").
                title:
                  type: string
      responses:
        '200':
          description: Updated spec summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/SpecSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /specs/{id}/lineage:
    get:
      operationId: getSpecLineage
      summary: Chain of custody for a spec
      description: >
        Walks `parent_spec_id` from the requested spec to the root library master,
        returning every hop in leaf-first order (`chain[0]` = the requested spec,
        last entry = the root).


        `behindBy` for each hop = the parent's current `contentVersion` minus this
        copy's `originVersion` (the version it was cloned from); null on the root
        because the root has no parent to drift behind.


        `originMeta` is the root's ingest provenance (`filename`, `sha256`, `loader`)
        when the root was file-ingested; null otherwise.
      tags: [specs]
      parameters:
        - $ref: '#/components/parameters/SpecId'
      responses:
        '200':
          description: Custody chain from requested spec to root
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/SpecLineage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /specs/{id}/style-source:
    post:
      operationId: setSpecStyleSource
      summary: Assign a style template to a spec (manual style-source pick)
      description: >
        Associates an existing style template (#30/#31) with the spec so the
        generator can render it with a real house style (#138). One template per
        spec; re-assigning replaces the previous template. `.SEC`/`.txt`/PDF
        masters carry no adopted visual style, so this is the narrowest useful
        slice of #125's style sources. Both the spec and the template must
        already exist — either missing → 404 (pre-checked, never via the FK error).
        A template assigned to any spec cannot be deleted (DELETE /templates/{id}
        → 409).
      tags: [specs]
      parameters:
        - $ref: '#/components/parameters/SpecId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId]
              properties:
                templateId:
                  type: string
                  format: uuid
                  description: Style template to associate with the spec.
      responses:
        '200':
          description: Style source assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/StyleSource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: clearSpecStyleSource
      summary: Clear a spec's style template (fall back to generator defaults)
      description: >
        Removes the spec→template association (#138). Idempotent — clearing an
        already-unset association is still 200. The spec must exist (unknown → 404).
      tags: [specs]
      parameters:
        - $ref: '#/components/parameters/SpecId'
      responses:
        '200':
          description: Style source cleared (styleSource is null)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        required: [styleSource]
                        properties:
                          styleSource:
                            type: 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /templates/{id}:
    get:
      operationId: getTemplate
      summary: Retrieve a style template with its rules
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: Template with rules
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      operationId: patchTemplate
      summary: Update template metadata (name and/or owner)
      description: >
        At least one of `name` or `owner` must be present. Pass `owner: null`
        to clear the owner field.
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  minLength: 1
                owner:
                  type: [string, 'null']
                  minLength: 1
                  description: Pass null to clear the owner
      responses:
        '200':
          description: Updated template metadata
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/TemplateMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteTemplate
      summary: Delete a style template and all its rules
      description: >
        A template assigned to any spec as its style source cannot be deleted
        (ON DELETE RESTRICT, #138) — the request returns 409 with the count of
        referencing specs. Clear the association on those specs first
        (DELETE /specs/{id}/style-source).
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '204':
          description: Template deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Template is in use by one or more specs as a style source and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    SpecId:
      name: id
      in: path
      required: true
      description: Spec UUID
      schema:
        type: string
        format: uuid
...
    TemplateId:
      name: id
      in: path
      required: true
      description: Template UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request — validation error or unsupported file type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict — a resource with that identity already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Unprocessable — file parsed but semantically unusable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string

    # ── Parse ──────────────────────────────────────────────────────────────────
...
    ParseWarning:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [root-continuation, empty-part, no-structure-found, unusual-part-count]
        lineHint:
          type: string
          description: 'Source location hint (e.g. "line 12: <text>")'
        suggestion:
          type: string

    # ── Specs ──────────────────────────────────────────────────────────────────
    SignalConflict:
      type: object
      required: [signal, reportedIlvl, reportedNodeType]
      properties:
        signal:
          type: integer
          enum: [1, 2, 3, 4, 5]
          description: The losing inference signal that disagreed with the winner
        reportedIlvl:
          type: integer
          description: Normalized ilvl the losing signal reported (part=0 … pr7=8)
        reportedNodeType:
          type: string
          enum: [spec, part, article, pr1, pr2, pr3, pr4, pr5, pr6, pr7, note, continuation]
    SourceTextSpan:
      type: array
      minItems: 2
      maxItems: 2
      prefixItems:
        - type: integer
          minimum: 0
        - type: integer
          minimum: 0
    SourceCommentFact:
      type: object
      required: [author, text, anchor]
      properties:
        author:
          type: string
        text:
          type: string
        anchor:
          $ref: '#/components/schemas/SourceTextSpan'
    SourceColorFact:
      type: object
      required: [color, coverage, spans]
      properties:
        color:
          type: string
        coverage:
          type: number
          minimum: 0
          maximum: 1
        spans:
          type: array
          items:
            $ref: '#/components/schemas/SourceTextSpan'
    SourceChoiceTokenFact:
      type: object
      required: [kind, options, span]
      properties:
        kind:
          type: string
          enum: [angle, bracket]
        options:
          type: array
          items:
            type: string
        span:
          $ref: '#/components/schemas/SourceTextSpan'
    SourceFacts:
      type: object
      description: >-
        Parser source facts for this paragraph. Present only when non-empty;
        unknown JSON keys are preserved for forward-compatible fact producers.
      additionalProperties: true
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/SourceCommentFact'
        colors:
          type: array
          items:
            $ref: '#/components/schemas/SourceColorFact'
        choiceTokens:
          type: array
          items:
            $ref: '#/components/schemas/SourceChoiceTokenFact'
        banner:
          type: string
        vanish:
          type: boolean
          const: true
    SpecNode:
      type: object
      required: [id, type, text, children, meta]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [spec, part, article, pr1, pr2, pr3, pr4, pr5, pr6, pr7, note, continuation]
        text:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/SpecNode'
        meta:
          type: object
          properties:
            vanish:
              type: boolean
              description: Hidden (w:vanish) paragraph — editorial note, not owner-facing
            source:
              type: string
              enum: [ufgs, arcat, cpi, unknown]
            revitParam:
              type: string
            baseVersion:
              type: integer
              minimum: 0
            conflicts:
              type: array
              description: >-
                DOCX inference signal disagreements for this paragraph. Present
                only when non-empty — the hierarchy level was ambiguous and the
                engine picked the highest-priority signal.
              items:
                $ref: '#/components/schemas/SignalConflict'
            sourceFacts:
              $ref: '#/components/schemas/SourceFacts'
    SpecTree:
      type: object
      required: [id, section, title, parts]
      properties:
        id:
          type: string
          format: uuid
        section:
          type: string
          description: 'CSI section number (e.g. "09 91 26"), or "unknown" if not detected'
        title:
          type: string
        parts:
          type: array
          items:
            $ref: '#/components/schemas/SpecNode'
        styleSource:
          description: >
            Manually assigned style template (#138), or null when none is set.
            Only present on GET /specs/{id}.
          oneOf:
            - $ref: '#/components/schemas/StyleSource'
            - type: 'null'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ParseWarning'
    StyleSource:
      type: object
      required: [templateId, templateName]
      properties:
        templateId:
          type: string
          format: uuid
        templateName:
          type: string
    SpecSummary:
      type: object
      required: [specId, title, section]
      properties:
        specId:
          type: string
          format: uuid
        title:
          type: string
        section:
          type: string
...
    LineageHop:
      type: object
      required: [specId, scope, name, contentVersion, originVersion, behindBy]
      description: >
        One node in the derivation chain. `originVersion` and `behindBy` are null
        on the root (it has no parent to drift behind).
      properties:
        specId:
          type: string
          format: uuid
        scope:
          type: string
          enum: [library, project]
          description: Whether this spec lives in a library or a project
        name:
          type: string
          description: Library or project name that owns this spec
        contentVersion:
          type: integer
          description: Current edit generation of this spec
        originVersion:
          type: [integer, 'null']
          description: Content version of the parent this spec was cloned from; null on root
        behindBy:
          type: [integer, 'null']
          description: >
            Parent's current contentVersion minus this copy's originVersion —
            how many parent edits this copy has not yet seen; null on root
    SpecLineage:
      type: object
      required: [chain, originMeta]
      properties:
        chain:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineageHop'
          description: >
            Derivation chain, leaf-first. `chain[0]` is the requested spec;
            the last entry is the root library master.
        originMeta:
          description: >
            Ingest provenance of the chain root (filename, sha256, loader);
            null when the root was not file-ingested.
          oneOf:
            - type: object
              required: [filename, sha256, loader]
              properties:
                filename:
                  type: string
                sha256:
                  type: string
                  minLength: 64
                  maxLength: 64
                loader:
                  type: string
            - type: 'null'
...
    StyleNodeType:
      type: string
      enum: [part, article, pr1, pr2, pr3, pr4, pr5, pr6, pr7]
      description: >
        CSI paragraph role. Maps to the nine generator levels: `part` (PART 1 –
        GENERAL), `article` (1.1 REFERENCES), `pr1`–`pr7` (A. / 1. / a. / 1) /
        a), then repeated 1) / a) paragraph tiers).
    StyleRule:
      type: object
      required: [nodeType, properties]
      properties:
        nodeType:
          $ref: '#/components/schemas/StyleNodeType'
        properties:
          $ref: '#/components/schemas/StyleProperties'
    StyleProperties:
      type: object
      description: >
        OOXML-derived style properties for a CSI paragraph role. Recognized keys
        are documented below; **unknown OOXML keys are preserved** via the JSONB
        catchall schema. All sub-objects also accept additional OOXML properties.
      properties:
        rPr:
          type: object
          description: Run (character) properties
          properties:
            rFonts:
              type: object
              description: Font references
              properties:
                ascii:
                  type: string
                hAnsi:
                  type: string
                cs:
                  type: string
                eastAsia:
                  type: string
              additionalProperties: true
            sz:
              type: integer
              description: Font size in half-points (e.g. 20 = 10 pt)
            b:
              type: boolean
              description: Bold
            i:
              type: boolean
              description: Italic
            caps:
              type: boolean
              description: All-caps
            smallCaps:
              type: boolean
            u:
              type: string
              description: Underline type (e.g. "single")
            strike:
              type: boolean
            color:
              type: string
              description: 'RRGGBB hex or "auto" (e.g. "FF0000")'
            highlight:
              type: string
          additionalProperties: true
        pPr:
          type: object
          description: Paragraph properties
          properties:
            spacing:
              type: object
              properties:
                before:
                  type: integer
                  description: Space before in twips
                after:
                  type: integer
                  description: Space after in twips
                line:
                  type: integer
                  description: Line spacing value in twips
                lineRule:
                  type: string
                  enum: [auto, exact, atLeast]
                contextualSpacing:
                  type: boolean
              additionalProperties: true
            ind:
              type: object
              description: Indentation (twips)
              properties:
                left:
                  type: integer
                right:
                  type: integer
                firstLine:
                  type: integer
                hanging:
                  type: integer
              additionalProperties: true
            jc:
              type: string
              enum: [left, center, right, both, distribute, start, end]
          additionalProperties: true
        numbering:
          type: object
          description: Numbering definition (from `numbering.xml` level)
          properties:
            ilvl:
              type: integer
            numFmt:
              type: string
            lvlText:
              type: string
            start:
              type: integer
          additionalProperties: true
      additionalProperties: true
    TemplateMeta:
      type: object
      required: [id, name, owner, createdAt]
      description: Template identity and metadata without style rules
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Unique template name
        owner:
          type: [string, 'null']
          description: Firm / owner label, or null if not provided
        createdAt:
          type: string
          format: date-time
    Template:
      allOf:
        - $ref: '#/components/schemas/TemplateMeta'
        - type: object
          required: [rules]
          properties:
            rules:
              type: array
              items:
                $ref: '#/components/schemas/StyleRule'
              description: One rule per styleable NodeType (9 max)
...